mwoauth looks like the fastest way to get oauth working.
This isn't available as a debian package, so we're installing directly.
Its requirements ARE available as Debian packages and have been
added to requirements.txt. One wrinkle: this specifies PyJWT > 1.0,
and only an earlier version is packaged. However, it looks like that
doesn't have any functional implications for mwoauth; 1.0 is a
security upgrade, but it's up to MediaWiki how they run their servers...