Add minimal expiring-soon page
This backs out the nonworking copy-paste of DurationField in favor of
a local install of the durationfield package (not available via
debian packages), and adds a page for admin to check up on apps likely
to expire soon, to facilitate autorenewal.

In future this page should 1) separate probably-expired from unexpired
apps, and 2) use the partner's Duration rather than a default.