Add time_open field to Application
We need to do a lot of computations around the time it takes to make
application decisions. In a better (read: Django 1.8+) world, we'd use
F expressions inside of aggregate/annotate to have the database do those
computations, but that feature isn't available to us. So let's
precompute and store the number of days that applications are open so we
can use it in database operations.