Account for possible nonuniqueness of WP usernames
The same username *may* represent different people on different
wikipedias (although it usually does not). We display homewiki for
disambiguation purposes and ensure that wp_usernames need not be
unique, but must be unique together with homewiki. We also require
homewiki - we will shortly be integrating it into the signup flow to
ensure that it is collected before Editor creation.