Update Finnish translation
This also exposed the issue that citation practices are different on the
Finnish wiki, so we can't interpolate the English tag as a variable;
the Finnish translator has no content to supply for that variable, but
the .po file won't compile without content. about.html has therefore
been rewritten to not require a variable.

This leads to a situation that's impossible to read (or write) if you
don't know HTML, so there's a translator comment. Developers and
translators will probably have to work together on this paragraph.