No, we didn't need that in save, I was right earlier
We'd moved the logic to the receiver, so it's all there. The issues with
some apps on the server not having had date_closed set must have been
fixed in an earlier commit; existing tests for that bug pass in this
version. Have set date_closed and days_open on the bad apps to match
the average for other closed apps.