Improve usability of coordinator designation interaction
I'm limiting the dropdown to actual coordinators so that admins don't
have to scroll through a giant list. I don't want to enforce/validate
this on the database level, though, as people may proceed through the
coordinator designation process in different orders, and also the exact
set of coordinators may change over time (and therefore the database
might be left with a previously-valid coordinator in an instance
record).