Fix bug whereby evaluating not_approved apps could cause server error
The if condition in the post_save turned out to be an elaborate way to
shoot myself in the foot with invisible type errors, so date_closed
and days_open were never being set, causing an error in template
rendering when get_num_days_open returned unexpected information.

Also adds some tests that should have existed ages ago and would
have caught this.