Document bug with, and fix for, deploying new languages with taggit
When you run python manage.py makemigrations on localhost, it creates
all the migrations you need for tags to work on localhost...but the
migrations for django-taggit are not in your version control since
taggit is a dependency. Then you deploy to the server, run python
manage.py migrate, and are sad, because your database schema does not
match your code. This is easy to fix! Once you realize you have to fix
it.