Make sure tags saved with blank fields don't fail
Normally, a blank CharField is saved as '', which means two blank
CharFields will violate a unique=True constraint. Instead, we need
CharFields to be nullable, because two null fields do NOT violate this
constraint.

This was an awful lot of mucking around in the guts of metaprogramming
for something that turned out to be a straightforward fix via an
option that was totally documented, but not apparently wreathed in the
neon snakes I required.